+++ includes/languages/english/modules/payment/emspay_tikkiepaymentrequest.php
@@ -1,0 +1,21 @@
+<?php
+
+define('MODULE_PAYMENT_EMSPAY_ORDER_DESCRIPTION', "Your order %s at %s");
+
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_TEXT_TITLE', "Tikkie Payment Request");
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_TEXT_DESCRIPTION', "Payment method is provided by EMS Online.");
+
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_DISPLAY_TITLE_TEXT', "Payment method display name");
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_DISPLAY_TITLE_DESCRIPTION', "Payment method name that will be displayed to the customer.");
+
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_STATUS_TEXT', "Enable Tikkie Payment Request");
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_STATUS_DESCRIPTION', "Do you want to enable Tikkie Payment Request?");
+
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_SORT_ORDER_TEXT', "Sort order for Tikkie Payment Request");
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_SORT_ORDER_DESCRIPTION', "The payment method with the lowest order is displayed first.");
+
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_ZONE_TEXT', "Payment Method availability Zone");
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_ZONE_DESCRIPTION', "If a zone is selected, then this payment method is enabled only for that zone.");
+
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_ERROR_TRANSACTION', "There was an error processing your payment. We apologize for the inconvenience. Please choose another payment method.");
+define('MODULE_PAYMENT_EMSPAY_TIKKIEPAYMENTREQUEST_ERROR_ALREADY_INSTALLED', "The Credit Card module is already installed.");