+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 *1.0.0*
 
 * Initial Release
+
+*1.0.1*
+
+* Changed library from ems-php to ginger-php
+* Renamed Sofort to Klarna Pay Now
+* Renamed Klarna to Klarna Pay Later
+* Add American Express, Tikkie Payment Request, WeChat